pyproject.toml
[project]
name = "jviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "quetzal-instruments>=0.33.7",
]
